+++ app/Views/core/header.php
@@ -1,5 +1,6 @@
 <?php
 $title ??= '';
+$url ??= '';
 ?>
 <!DOCTYPE html>
 <html lang="<?= \Config\Services::request()->getLocale() ?>">
@@ -19,9 +20,13 @@ $title ??= '';
             gtag('config', '<?= $_ENV['analytics_google'] ?>');
         </script>
     <?php } ?>
+    <?php if (($title !== 'Welcome' && $url === '') || $_ENV['app_baseEnvironmentProjectUrl'] !== $_ENV['app_baseCanonicalProjectUrl']) { ?>
+    <meta name="robots" content="noindex">
+    <?php } ?>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta charset="UTF-8">
     <title><?= (isset($pageTitle) ? $pageTitle . ' | ' : '') . (isset($pageTitleType) ? $pageTitleType . ' | ' : '') . $title ?> | <?= lang('Template.projectName') ?></title>
+    <link rel="canonical" href="<?= $_ENV['app_baseCanonicalProjectUrl'] ?>/<?= \Config\Services::request()->getLocale() ?>/<?= $url ?>">
     <link rel="preload" href="/asset/font/lora-regular.woff2" as="font" type="font/woff2" crossorigin="anonymous">
     <link rel="preload" href="/asset/font/lora-semibold.woff2" as="font" type="font/woff2" crossorigin="anonymous">
     <link rel="preload" href="/asset/font/lora-italic.woff2" as="font" type="font/woff2" crossorigin="anonymous">